pyproject.toml
[project]
name = "airtfast"
version = "1.0.0"
description = "Real-time financial analytics platform for telecom vendors."
authors = [
  { name = "Ash", email = "you@example.com" }
]
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
  "Flask==2.3.3",
  "Flask-Login==0.6.2",
  "Flask-Migrate==4.0.5",
  "Flask-SQLAlchemy==3.1.1",
  "Flask-WTF==1.1.1",
  "flask-restx==1.3.0",
  "Flask-APScheduler==1.13.1",
  "psycopg2-binary==2.9.10",
  "python-decouple==3.8",
  "python-dotenv==1.1.0",
  "Werkzeug==2.3.8",
  "gunicorn==21.2.0",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["apps"]

[tool.setuptools.package-dir]
apps = "apps"
